small warning fixlet 2005-12-25  Dalibor Topic  <robilad@kaffe.org>
        * kaffe/kaffevm/systems/unix-jthreads/jthread.c:
        Cast jlong to long in debug statement to fix a compiler warning.

+++ kaffe/kaffevm/systems/unix-jthreads/jthread.c
@@ -2272,7 +2272,7 @@ retry:
 		} else {
 			maxWait = firstAlarm - curTime;
 		}
-		DBG(JTHREADDETAIL, dprintf("handleIO(sleep=%d) maxWait=%d\n", canSleep, maxWait); );
+		DBG(JTHREADDETAIL, dprintf("handleIO(sleep=%d) maxWait=%ld\n", canSleep, (long) maxWait); );
 	}
 
 #if USE_POLL
